ArgMax Function (#22)
* ArgMax function

The argmax is a commonly used data wrangling operation. Efficiently implementing it from scratch is a bit non-trivial. Therefore, we want to provide this functionality as a built-in feature of nada-numpy

Co-Authored-By: rostin79s <121708191+rostin79s@users.noreply.github.com>
Co-Authored-By: lfolkerts <6434841+lfolkerts@users.noreply.github.com>

* Fix yaml files

In arg_max.py, the `current_indx` is set to `Integer(0)` and `one` is set to `Integer(1)`

Removed the `current_indx` and `public_one` from the yaml files.

Added a new line at the end of nada_project.toml file.

* Fix

Appended newlines
Renaming variables

Co-Authored-By: lfolkerts <6434841+lfolkerts@users.noreply.github.com>
Co-Authored-By: rostin79s <121708191+rostin79s@users.noreply.github.com>

* Removed the result index from input

`result` is initialized to -1 instead of reading it as input.
The condition is changed to be `>=` instead of `>`.
Removed the `index` input from the test files.

Co-Authored-By: lfolkerts <6434841+lfolkerts@users.noreply.github.com>
Co-Authored-By: rostin79s <121708191+rostin79s@users.noreply.github.com>

* Updated function header

Co-Authored-By: lfolkerts <6434841+lfolkerts@users.noreply.github.com>
Co-Authored-By: rostin79s <121708191+rostin79s@users.noreply.github.com>

---------

Co-authored-by: rostin79s <121708191+rostin79s@users.noreply.github.com>
Co-authored-by: lfolkerts <6434841+lfolkerts@users.noreply.github.com>
Co-authored-by: Dimitris Mouris <dimitris@nillion.com>